Crate rtx_core

The Core of rtx - roughly the equivalent of TeX conversion.

Re-exports

pub use crate::common::error::*;

Modules

aux_macros

Auxiliary macros

binding

The programmable API foundation for creating bindings of LaTeX sty/cls libraries

comment

TeX comments as standalone objects

common

Common abstractions that are useful at various stages of Core processing

definition

All possible definitions for TeX-native commands (expandable, primitive, constructor,…)

document

An abstraction layer over the converted XML document

gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

keyval

A LaTeX-like Key-Value object

keyvals

A collection of Key-Value objects, typically from a single LaTeX argument

ligature

Rules for combining together characters and other text rules

list

A list of Digested objects

mouth

The mouth is a thin interface over a file, responsible for reading characters and associating them with catcodes

parameter

The abstraction layer used by the Gullet to read arguments for the various kinds of TeX object definitions

rewrite

Rules for rewriting the constructed XML document, after core processing has completed

state

A global, singleton, mutable State - hosts almost all TeX-facing runtime information for the conversion

stomach

The stomach is an abstraction responsible for digesting Tokens and Registers prepared by the Gullet into Boxes

tbox

A TeX-like digested Box

token

The Token and its Catcode

tokens

Grouping Tokens together

util

Auxilary utilities that do not participate in the main conversion abstraction

whatsit

A TeX-like digested Whatsit

Macros

AssignRegister

Debug

Error

Explode

Explode a string into a list of tokens, all w/catcode OTHER (except space).

ExplodeText

Similar to Explode, but convert letters to catcode LETTER and others to OTHER Hopefully, this is essentially correct WITHOUT resorting to catcode lookup?

Fatal

Font

FontDirective

Info

NewDefault

NewDefaultV

Number

Roman

Convert the number to upper case roman numerals, returning a list of LaTeXML::Core::Token

SetCounter

T_ACTIVE

macro for an ACTIVE char token

T_ALIGN

macro for an ALIGN “&” token

T_ARG

macro for a numbered ARG token

T_BEGIN

macro for a BEGIN “{” token

T_COMMENT

macro for a COMMENT content token

T_CR

macro for a CR “\n” token

T_CS

macro for a command sequence token

T_END

macro for an END “}” token

T_LETTER

macro for a LETTER token

T_MARKER

macro for a tracing MARKER token

T_MATH

macro for a MATH “$” token

T_OTHER

macro for an OTHER code token

T_PARAM

macro for a PARAM “#” token

T_SMUGGLE_THE

macro for a SmuggleThe token (see gullet::invoke_and_read_x_token)

T_SPACE

macro for a SPACE “ “ token

T_SUB

macro for a SUB “_” token

T_SUPER

macro for a SUPER “^” token

Token

Token constructor macro (defaults to OTHER code)

Tokens

convenience macro for assembling a Tokens object from different pieces (Token, Vec<Token>, Tokens)

Warn

after_digest

after_digest_simple

after_digest_single

before_digest

before_digest_simple

before_digest_single

construct

count_unpack_opt

count_unpack_ref

count_unpack_to_string

fatal

fontmap

Instantiates a Font, using the Font fields as keys, and calling .into() for each value. The specification can be partial - missing fields are taken via the Default trait.

forbidMath

generate_message

getter

map

A flexary macro for constructing HashMap<String, T> maps, where T is generic.

mixvec

A variant on vec! where each argument receives an additional .into() call best used with an outer context that explicitly provides the expected type, such as

noprimitive

noreplacement

primitiveproc

primitivesub

print_stderr

println_stderr

prop_bool

prop_digested

prop_str

prop_string

prop_whatsit

properties

raw_char_map

A flexary macro for constructing HashMap<char, T> maps, where T is generic

raw_map

A flexary macro for constructing HashMap<K, V> maps, where K and V are both generic (inferred at time of use)

reader

reader_predigest

replacement

requireMath

reversion

reversion_digested

rewrite_replace_sub

roman

Convert the number to lower case roman numerals, returning a list of LaTeXML::Core::Token

s

The s! macro is a briefer alias for format!, otherwise completely identical

set

Simple generic helper for HashSet<K,V> creation Source: https://riptutorial.com/rust/example/4149/create-a-hashset-macro

setter

sizersub

static_map

! Helper macros for quicker and more idiomatic construction and access to data structures. A flexary macro for constructing HashMap<&'static str, &'static str> maps

stored_map

A flexary macro for constructing HashMap<String, Stored> maps

string_map

A flexary macro for constructing HashMap<String, String> maps

tagsub

transfer_default

transfer_opt_default

undigested

unpack

unpack_opt

An alternative to unpack!(args=>arg1,arg2...) which allows for optional arguments.

unpack_opt_ref

Meant to be used for unpacking &Vec<Option> in particular.

unpack_to_number

unpack_to_string

unpacks a Vec<ArgWrap> argument into a flexary number of String-typed variables uses the usual .to_string() method from the Display trait.

unpack_to_token

unref

Convenience macro to flexibly unpack a collection of Vec<ArgWrap> arguments into individual Tokens variables.

Structs

Core

The Core conversion runtime

CoreOptions

Configuration for the Core processing

Digested

An Arc-guarded abstraction for any object encountered at the “digested” phase of processing

Enums

DigestedData

These are all kinds of data which we consider officially supported as outputs from the digestion phase of TeX, i.e. from invoking a token.

TexMode

The current TeX processing mode

Traits

BoxOps

Common operations for Box-like (digested) data