Module rtx_core::alignment

Support for TeX-like Alignments

Representation of aligned structures

An “Alignment” is an array/tabular construct as:
… or, for math mode … (where initially, each XMCell will contain an XMArg to indicate individual parsing of each cell’s content is desired)

An Alignment object is a sort of fake Whatsit; It takes some magic to sneak it into the Digestion stream (see TeX.pool @start@alignment), but it needs to be created BEFORE the contents of the alignment are digested, since we stuff a lot of information into it (row, column boxes, borders, spacing, etc…) But once it has been captured, it should otherwise act like a Whatsit and be responsible for construction (be_absorbed), and sizing estimation (computeSize)

Ultimately, this should be better tied into DefConstructor because an Alignment currently doesn’t know what CS created it (debugging!); Also, it would better connect the things being constructed, reversion, etc.

Modules

cell

template

Support for tabular/array environments

Structs

Alignment

AlignmentConfig

Functions

matrix_template

parse_alignment_template

read_alignment_template

a reader for the Template parameter type

Type Aliases

CloseColumnFn

CloseContainerFn

CloseRowFn

OpenColumnFn

OpenContainerFn

OpenRowFn