Struct rtx_core::alignment::Alignment

pub struct Alignment { /* private fields */ }

Implementations

impl Alignment

pub fn new(config: AlignmentConfig) -> Self

Create a new Alignment. config can contain: template : an Alignment::Template object openContainer = sub($doc,%attrib); creates the container element with given attributes closeContainer = sub($doc); closes the container openRow = sub($doc,%attrib); creates the row element with given attributes closeRow = closes the row openColumn = sub($doc,%attrib); creates the column element with given attributes closeColumn = closes the column properties = hashref containing extra attributes for the container element.

pub fn get_template(&self) -> &Template

pub fn current_row(&self) -> Option<&Row>

pub fn current_row_mut(&mut self) -> Option<&mut Row>

pub fn new_row(&mut self) -> Option<&Row>

pub fn remove_row(&mut self) -> Option<Row>

pub fn prepend_rows(&mut self, new_rows: Vec<Row>)

pub fn append_rows(&mut self, new_rows: Vec<Row>)

pub fn rows(&self) -> &VecDeque<Row>

pub fn add_line(&mut self, border: &str, cols: Vec<usize>)

pub fn next_column(&mut self) -> Option<&mut Cell>

pub fn last_column(&mut self) -> Option<&mut Cell>

pub fn current_column_number(&self) -> usize

pub fn current_row_number(&self) -> usize

pub fn current_column(&mut self) -> Option<&mut Cell>

pub fn get_column(&mut self, n: usize) -> Option<&mut Cell>

pub fn get_column_mut(&mut self, n: usize) -> Option<&mut Cell>

pub fn add_before_row(&mut self, boxes: Vec<Digested>)

pub fn add_after_row(&mut self, boxes: Vec<Digested>)

pub fn omit_column(&mut self)

pub fn omit_next_column(&mut self)

pub fn get_column_before(&mut self) -> Tokens

pub fn get_column_after(&mut self) -> Tokens

pub fn revert(&self, _state: &State) -> Result<Tokens>

pub fn set_reversion(&mut self, rev: Tokens)

pub fn set_content_reversion(&mut self, rev: Tokens)

pub fn is_in_row(&self) -> bool

pub fn is_in_column(&self) -> bool

pub fn start_row( &mut self, pseudorow: bool, stomach: &mut Stomach, state: &mut State ) -> Result<()>

pub fn end_row(&mut self, stomach: &mut Stomach, state: &mut State) -> Result<()>

pub fn start_column( &mut self, pseudorow: bool, stomach: &mut Stomach, state: &mut State ) -> Result<()>

pub fn end_column( &mut self, stomach: &mut Stomach, state: &mut State ) -> Result<()>

pub fn set_in_tabular_head(&mut self)

pub fn unset_in_tabular_head(&mut self)

pub fn is_in_tabular_head(&self) -> bool

pub fn normalize_alignment(&mut self, state: &mut State) -> Result<()>

====================================================================== Normalize an alignment before construction

consolodating column & row spanning information

scanning for empty rows & columns and collapsing them (while accounting for spanning, and copying borders appropriately) Note that a trailing \ in allignment (often needed to effect \hline) causes an empty row at the end. Other fancy layout fine-tuning often involves adding extra rows & columsn for spacing. HTML’s table model is more forgiving that TeX’s, so we don’t need these extras and, in fact, they often mess up the html layout! However, math alignments, and those with expected structure (eg. eqnarray) should generally NOT have rows & columns collapsed — except the last row!

Also note the inconsistency between TeX & HTML’s table models regarding spans. \multicolumn creates a cell that covers a certain number of columns which are then omitted from the LaTeX AND the HTML. OTOH, \multirow creates a cell which overlaps following rows! The & is still needed to allocate the cells in those rows. And in fact they need not even be empty! TeX will just pile them up! However, in HTML the spanned rows ARE omitted!

pub fn normalize_cell_sizes(&mut self, state: &mut State) -> Result<()>

Compute (approximate) sizes of all cells

pub fn normalize_mark_spans(&mut self) -> Result<()>

Mark any cells that are covered by rowspan or colspan

pub fn normalize_prune_rows(&mut self) -> Result<()>

Scan for and remove empty rows but copying borders and adjusting rowspan’s & colspan’s appropriately.

pub fn normalize_prune_columns(&mut self) -> Result<()>

Scan for and remove empty columns but copying borders and adjusting rowspan’s & colspan’s appropriately.

pub fn normalize_sum_sizes(&mut self) -> Result<()>

pub fn get_properties_mut(&mut self) -> &mut HashMap<String, Stored>