Module rtx_core::binding::content

Structs

FindFileOptions

InputDefinitionOptions

InputOptions

RequireOptions

Functions

build_invocation

convert_latex_args

Convert a LaTeX-style argument spec to our Package form. Ie. given $nargs and $optional, being the two optional arguments to something like \newcommand, convert it to the form we use

digest_if

digest_literal

digest_text

do_expand

find_file

find_file_aux

input

input_content

input_definitions

TODO: Flesh out with the full infrastructure, incremental functionality for now.

install_tag

load_binding

load_class

load_external_binding

load_font_map

load_tex_content

merge_font

preload_font_map

process_options

require_package

This (and FindFile) needs to evolve a bit to support reading raw .sty (.def, etc) files from the standard texmf directories. Maybe even use kpsewhich itself (INSTEAD of pathname_find ???) Another potentially useful option might be that if we are reading a raw file, perhaps it should just get digested immediately, since it shouldn’t contribute any boxes.

require_resource

select_relaxng_schema