Function rtx_core::binding::counter::dialect::ref_step_counter

pub fn ref_step_counter(
    ctype: &str,
    noreset: bool,
    stomach: &mut Stomach,
    state: &mut State
) -> Result<HashMap<String, Stored>>

Analog of \refstepcounter, steps the counter and returns a hash containing the keys refnum and id. This makes it suitable for use in a properties option to constructors. The id is generated in parallel with the reference number to assist debugging.