Module rtx_core::common::arena

An arena for interning strings (global, mutable, single thread only)

Note: works under the assumption of a single threaded, short-lived process, where memory starvation will not be an issue. It would need a hard reset if the same process does multiple conversions, and a different implementation if one needs a thread-local arena.

Statics

ANY_SYM

the unique symbol for str value “ANY”

EMPTY_SYM

the unique symbol for the empty str value “”

GLOBAL_DEFS_SYM

the unique symbol for str value “\globaldefs”

LTX_P_SYM

the unique symbol for str value “ltx:p”

LTX_STAR_SYM

the unique symbol for str value “ltx:*”

PCDATA_SYM

the unique symbol for str value “#PCDATA”

Functions

as_static

Pin-and-resolve a string into the arena, obtaining a &'static str reference to it (useful for str lifetime collisions)

pin

Assign a string into the arena, returning a unique symbol associated with it

resolve

Resolve the data associated with a unique symbol from this arena, asserting it’s already present

try_resolve

Attempt to resolve a string associated with a unique symbol from this arena, None if missing