Module rtx_core::common::numeric_ops

Constants

EPSILON

ROUNDING_HALF

SCALES

UNITY

UNITY_F64

Traits

NumericOps

Functions

fixpoint

Convert float to a fixed-point number If unit is given, it is number of units PER SCALED-POINT! (hence, extra division) AND, note that the float is rounded and THEN truncated after multiplying by units! to mimic TeX’s behavior.

kround

An attempt at rounding floats to integers (like scaled points), in a (hopefully) Knuthian manner (like round_decimals \S102 in Tex The Program)

round_to

Round $number to $prec decimals (0…6) attempting to do so portably.