Struct rtx_core::definition::register::Register

pub struct Register {
    pub cs: Token,
    pub address: String,
    pub parameters: Option<Parameters>,
    pub register_type: RegisterType,
    pub readonly: bool,
    pub value: Option<RegisterValue>,
    pub getter: Option<RegisterGetterClosure>,
    pub setter: Option<RegisterSetterClosure>,
    pub default: Option<RegisterValue>,
    pub mathglyph: Option<char>,
    pub role: Option<SymbolU32>,
    pub locator: Locator,
}

A struct representing a TeX register

Fields
cs: Token

the public command sequence for this register

address: String

the internal address for this register

parameters: Option<Parameters>

associated parameters, if any

register_type: RegisterType

the type of values accepted by this register (Number, Dimension, …)

readonly: bool

read-only flag (default: false)

value: Option<RegisterValue>

the current value

getter: Option<RegisterGetterClosure>

reader for a value

setter: Option<RegisterSetterClosure>

setter for a value

default: Option<RegisterValue>

a default value

mathglyph: Option<char>

the unicode corresponding to the \mathchar of value (for chardef)

role: Option<SymbolU32>

TODO: what does this do for CharDef ?

locator: Locator

the source point of origin for this register definition

Implementations

impl Register

pub fn is_readonly(&self) -> bool

checks the readonly flag

pub fn new_chardef( cs: Token, value: Option<RegisterValue>, mathglyph: Option<char>, role: Option<SymbolU32> ) -> Self

A CharDef is a specialized register; You can’t assign it; when you invoke the control sequence, it returns the result of evaluating the character (more like a regular primitive). When mathglyph is provided, it is the unicode corresponding to the \mathchar of value

pub fn get_address(&self) -> Cow<'_, str>