Struct rtx_core::document::Document

pub struct Document {
    pub document: XmlDoc,
    pub pending: Vec<Node>,
    pub node: Node,
    pub node_boxes: HashMap<usize, Digested>,
    pub node_fonts: HashMap<u64, Font>,
    pub idstore: HashMap<String, Node>,
    pub rewrite_labels: HashMap<String, String>,
    /* private fields */
}

Fields
document: XmlDocpending: Vec<Node>node: Nodenode_boxes: HashMap<usize, Digested>node_fonts: HashMap<u64, Font>idstore: HashMap<String, Node>rewrite_labels: HashMap<String, String>
Implementations

impl Document

pub fn new() -> Self

pub fn get_element(&self) -> Option<Node>

Get the element at (or containing) the current insertion point.

pub fn findnodes(&mut self, xpath: &str, node_opt: Option<&Node>) -> Vec<Node>

Find the nodes according to the given xpath expression, the xpath is relative to $node (if given), otherwise to the document node.

pub fn get_xpath(&mut self) -> &mut XPath

Get an XPath context that knows about our namespace mappings.

pub fn findnode(&mut self, xpath: &str, node: Option<&Node>) -> Option<Node>

Like findnodes, but only returns the first matched node

pub fn findvalues( &mut self, xpath: &str, node_opt: Option<&Node> ) -> Vec<String>

Like findnodes, but expects an xpath that evaluates to a literal value (e.g. for attributes)

pub fn get_node(&self) -> &Node

pub fn get_node_mut(&mut self) -> &mut Node

pub fn get_document(&self) -> &XmlDoc

pub fn get_document_mut(&mut self) -> &mut XmlDoc

pub fn finalize(&mut self) -> Result<()>

pub fn absorb( &mut self, object: &Digested, props_opt: Option<HashMap<String, Stored>> ) -> Result<()>

Document construction at the Current Insertion Point.

absorb the given $box into the DOM (called from constructors). This will return a list of whatever nodes were created. Note that this may include nodes that are children of other nodes in the list or nodes that are no longer in the document. Also, note that when a text nodes is appended to, the complete text node is in the list, not just the portion that was added. [Note that recording the nodes being constructed isn’t all that costly, but filtering them for parent/child relations IS, particularly since it usually isn’t needed]

A box that is a TBox, or List, or Whatsit, is responsible for carrying out its own insertion, but it should ultimately call methods of Document that will record the nodes that were created. $box can also be a plain string (Digested::Postponed) which will be inserted according to whatever font, mode, etc, are in %props.

pub fn get_constructed_nodes(&self) -> &[Node]

pub fn absorb_string( &mut self, object: &str, props: &HashMap<String, Stored> ) -> Result<Option<Node>>

This is a refactored else cases from the main absorb routine, to allow for better type hygiene

pub fn insert_element( &mut self, qname: &str, content: Vec<&Digested>, attrib: Option<HashMap<String, String>> ) -> Result<Node>

Shorthand for open,absorb,close, but returns the new node.

pub fn insert_pi( &mut self, op: &str, attributes_opt: Option<HashMap<String, String>> ) -> Result<()>

Insert a ProcessingInstruction of the form Does NOT move the current insertion point to the PI, but may move up past a text node.

pub fn open_element( &mut self, qname: &str, attributes: Option<HashMap<String, String>>, font_opt: Option<&Font> ) -> Result<Node>

pub fn close_element(&mut self, qname: &str) -> Result<Option<Node>>

Note: This closes the deepest open node of a given type. This can cause problems with auto-opened nodes, esp. ones for fontswitches! Since this is an “explicit request”, we’re currently skipping over those nodes, ie. we’re automatically closing them, even if they’re the same type as we’re asking to close!!! This is kinda risky! Maybe we should try to request closing of specific nodes.

pub fn is_openable(&self, qname: &str) -> bool

pub fn is_closeable<T: IntoVDQS>(&self, tags: T) -> Option<Node>

Check whether it is possible to close each element in @tags, any intervening nodes must be autocloseable. returning the last Some(node) that would be closed if it is possible, otherwise None

pub fn maybe_close_element(&mut self, qname: &str) -> Result<Option<Node>>

pub fn close_to_node(&mut self, node: &Node, _ifopen: bool) -> Result<()>

Closes all nodes until $node becomes the current point.

pub fn close_node(&mut self, node: &Node) -> Result<()>

Closes all nodes until $node is closed.

pub fn maybe_close_node(&mut self, node: &Node) -> Result<()>

Only if needed/possible: closes all nodes until $node is closed

pub fn close_node_with_strictness( &mut self, strict: bool, node: &Node ) -> Result<()>

pub fn get_tag_action_list( &self, tag: SymbolU32, when: TagOptionName ) -> Vec<TagConstructionClosure>

get the actions that should be performed on afterOpen or afterClose

pub fn serialize_to_string(&self) -> String

pub fn serialize_aux( &self, node: &Node, depth: usize, noindent: bool, heuristic: bool ) -> String

We ought to try for something close to C14N (http://www.w3.org/TR/xml-c14n), but keep XML declaration, comments and don’t convert empty elements.

pub fn set_node(&mut self, node: &Node)

pub fn insert_math_token( &mut self, text: &str, attributes: HashMap<String, String>, font_opt: Option<&Font> ) -> Result<Node>

pub fn insert_comment(&mut self, text: &str) -> Result<Node>

Insert a new comment, or append to previous comment. Does NOT move the current insertion point to the Comment, but may move up past a text node.

pub fn open_text(&mut self, text: &str, font: &Font) -> Result<Option<Node>>

pub fn close_text_internal(&mut self) -> Result<Node>

pub fn close_node_internal(&mut self, node: &Node) -> Result<()>

Close node, and any current nodes below it. No checking! Use this when you’ve already verified that node can be closed. and, of course, node must be current or some ancestor of it!!!

pub fn merge_attributes( &mut self, from: Node, to: &mut Node, _force: Option<HashSet<&'static str>> ) -> Result<()>

pub fn open_text_internal(&mut self, text: &str) -> Result<Node>

pub fn record_constructed_node(&mut self, node: &Node)

Note that a box has been absorbed creating node; This does book keeping so that we can return the sequence of nodes that were added by absorbing material.

pub fn filter_deletions(&self, nodes: Vec<Node>) -> Vec<Node>

pub fn filter_children(&self, nodes: Vec<Node>) -> Vec<Node>

Given a list of nodes such as from ->absorb, filter out all the nodes that are children of other nodes in the list.

pub fn get_insertion_context(&self, levels_opt: Option<usize>) -> Result<String>

Return a string indicating the path to the current insertion point in the document. if $levels is defined, show only that many levels

pub fn find_insertion_point( &mut self, qname: &str, has_opened_opt: Option<SymbolU32> ) -> Result<Node>

Find the node where an element with qualified name $qname can be inserted. This will move up the tree (closing auto-closable elements), or down (inserting auto-openable elements), as needed.

pub fn node_set_attribute(&mut self, key: &str, value: &str) -> Result<()>

pub fn node_get_attribute(&mut self, name: &str) -> Option<String>

pub fn set_attribute( &mut self, node: &mut Node, key: &str, value: &str ) -> Result<()>

pub fn add_ss_values( &mut self, node: &mut Node, key: &str, values_str: &str ) -> Result<()>

pub fn add_class(&mut self, node: &mut Node, class: &str) -> Result<()>

pub fn unrecord_id(&mut self, id: &str)

pub fn record_node_ids(&mut self, node: &Node) -> Result<()>

These are used to record or unrecord, in bulk, all the ids within a node (tree).

pub fn unrecord_node_ids(&mut self, node: &Node)

pub fn modify_id(&mut self, id: String) -> String

Get a new, related, but unique id Sneaky option: try “ID_SUFFIX” as a suffix for id, first.

pub fn lookup_id(&self, id: &str) -> Option<&Node>

pub fn mark_xmnode_visibility(&mut self) -> Result<()>

pub fn set_node_box(&mut self, node: &Node, digested: Digested)

Record the Box that created this node.

pub fn get_node_box(&self, node: &Node) -> Option<Digested>

pub fn set_node_font(&mut self, node: &mut Node, font: &Font) -> Result<()>

Record the Font of a node

pub fn copy_node_font(&mut self, from: &Node, to: &mut Node) -> Result<()>

pub fn merge_node_font_rec(&mut self, node: &Node, font: &Font) -> Result<()>

Possibly a sign of a design flaw; Set the node’s font & all children that HAD the same font.

pub fn set_box_font(&mut self, node: &mut Node) -> Result<()>

pub fn get_node_font(&self, node: &Node) -> &Font

pub fn has_node_font(&self, node: &Node) -> bool

pub fn get_node_language(&self, node: &Node) -> String

pub fn remove_node(&mut self, node: &mut Node)

pub fn open_element_at( &mut self, point: &mut Node, qname: &str, attributes: Option<HashMap<String, String>>, font_opt: Option<Font> ) -> Result<Node>

This opens a new element at the specified point, rather than the current insertion point. This is useful during document rearrangement or augmentation that may be needed later in the process.

pub fn close_element_at(&mut self, node: &mut Node) -> Result<()>

Whenever a node has been created using openElementAt, closeElementAt ought to be used to close it, when you’re finished inserting into $node. Basically, this just runs any afterClose operations.

pub fn after_open(&mut self, node: &mut Node) -> Result<()>

pub fn after_close(&mut self, node: &mut Node) -> Result<()>

pub fn append_clone( &mut self, node: &mut Node, new_children: Vec<Node> ) -> Result<()>

pub fn wrap_nodes( &mut self, qname: &str, nodes: Vec<Node> ) -> Result<Option<Node>>

pub fn unwrap_nodes(&mut self, node: Node) -> Result<Node>

Unwrap the children of $node, by replacing $node by its children.

pub fn replace_node(&mut self, node: Node, with: Vec<Node>) -> Result<Node>

pub fn rename_node(&mut self, _node: &mut Node, _newname: &str) -> Result<Node>

pub fn trim_node_whitespace(&mut self, node: &mut Node) -> Result<()>

pub fn add_resource(&mut self, resource: Resource) -> Result<()>

pub fn process_pending_resources(&mut self) -> Result<()>

pub fn make_error(&mut self, error_class: &str, _content: &str) -> Result<()>

pub fn float_to_element( &mut self, qname: &str, closeifpossible: bool ) -> Result<Option<Node>>

Find a node in the document that can contain an element qname

pub fn float_to_label(&mut self) -> Option<Node>

pub fn set_box_to_absorb(&mut self, arg: Option<Digested>)

pub fn expire_box_to_absorb(&mut self)

pub fn load_labels_for_rewrite(&mut self) -> Result<()>

pub fn generate_id(&mut self, node: &mut Node, prefix: &str) -> Result<()>

This function computes an xml:id for a node, if it hasn’t already got one. It is suitable for use in Tag afterOpen as Tag('ltx:para',afterOpen=>sub { GenerateID(@_,'p'); }); It generates an id of the form . The parent node (the one with ID=) also maintains a counter stored in an attribute _ID_counter_<prefix> recording the last used for amongst its descendents.

pub fn replace_tree( &mut self, new: Node, old: &mut Node ) -> Result<Option<Node>>

pub fn append_tree(&mut self, node: &mut Node, data: Vec<Node>) -> Result<()>