Module rtx_core::gullet

The Gullet is responsible for reading Tokens and other data from the Mouth

Structs

Gullet

MouthRuntime

Statics

GULLET

Functions

close_mouth

do_expand

flush

Obscure, but the only way I can think of to End!! (see \bye or \end{document}) Flush all sources (close all pending mouth’s)

flush_mouth

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds to TeX’s \endinput

get_location

get_locator

get_pending_comment

has_more_input

if_next

initialize_gullet

is_column_end

open_mouth

read_arg

read_balanced

read_balanced approximates TeX’s scan_toks (but doesn’t parse \def parameter lists) and only optionally requires the openning “{”. It may return comments in the token lists. it optionally (do_expand) expands while reading, but deferring \the and related. The is_macrodef flag affects whether # parameters are “packed” for macro bodies. If require_open is true, the opening T_BEGIN has not yet been read, and is required.

read_dimension

read_float

====================================================================== Float, a floating point number. Similar to factor, but does NOT accept comma! This is NOT part of TeX, but is convenient.

read_glue

read_keyword

Match the input against a set of keywords; Similar to readMatch, but the keywords are strings, and Case and catcodes are ignored; additionally, leading spaces are skipped. AND, macros are expanded.

read_match

Match the input against one of the Token or Tokens in @choices; return the matching one or undef.

read_mu_dimension

read_mu_glue

read_mu_unit

read_next_conditional

reads and discards tokens, until it encounters a conditional, if any

read_non_space

Read a single non-space token

read_normal_integer

= | | ’ | “ | `

read_number

====================================================================== Integer, Number

read_optional

Read and return a LaTeX optional argument; returns C<$default> if there is no ‘[’, otherwise the contents of the []. Note that this returns an empty array if [] is present, i.e. “[contents]” in TeX will lead to Tokens(contents), otherwise returns None

read_raw_line

Read the next raw line (string); primarily to read from the Mouth, but keep any unread input!

read_register_value

read_rubber

read_token

read_tokens_value

Apparent behaviour of a token value (ie \toks#=)

read_until

Return a (balanced) sequence tokens until a match against one of the Tokens in @delims. Note that Braces on input hides the contents from matching, so this assumes there wont be braces in $delim! But, see readUntilBrace for that case.

read_until_brace

reads until it encounters a Catcode::BEGIN token

read_until_token

Convenience method wrapping around read_until TODO: This seems to be the wrong Rust type interface, we need to rework…

read_value

read_x_non_space

Read a single expanded, non-space, token

read_x_token

Read the next non-expandable token (expanding tokens until there’s a non-expandable one). Note that most tokens pass through here, so be Fast & Clean! readToken is folded in. Toplevel' processing, (if toplevelis true), used at the toplevel processing by Stomach, will step to the next input stream (Mouth) if one is available,toplevel` is doing TWO distinct things. When true:

reading_from_mouth

Do something, while reading tokens from a specific Mouth. This reads ONLY from that mouth (or any mouth openned by code in that source), and the mouth should end up empty afterwards, and only be closed here.

skip_one_space

skip_spaces

unread

unread_mut

same as unread, but drains the tokens from its contents

unread_one

same as unread, but only for a single Token

unread_vec

same as unread, but does not require the Tokens wrapper

with_mouth_mut