Struct rtx_core::gullet::Gullet

pub struct Gullet {
    pub mouth: Option<MouthRuntime>,
    pub mouthstack: VecDeque<MouthRuntime>,
    pub pending_comments: VecDeque<Token>,
    /* private fields */
}

Fields
mouth: Option<MouthRuntime>mouthstack: VecDeque<MouthRuntime>pending_comments: VecDeque<Token>
Implementations

impl Gullet

pub fn flush_mouth(&mut self, state: &mut State)

This flushes a mouth so that it will be automatically closed, next time it’s read Corresponds (I think) to TeX’s \endinput

pub fn flush(&mut self, state: &mut State)

Obscure, but the only way I can think of to End!! (see \bye or \end{document}) Flush all sources (close all pending mouth’s)

pub fn has_more_input(&mut self) -> bool

pub fn open_mouth(&mut self, mouth: Mouth, autoclose: bool)

pub fn close_mouth(&mut self, forced: bool, state: &mut State)

pub fn get_mouth(&self) -> Option<&Mouth>

pub fn get_mouth_mut(&mut self) -> Option<&mut Mouth>

pub fn read_token(&mut self, state: &mut State) -> Option<Token>

Low-level readers: read token, read expanded token

Note that every char (token) comes through here (maybe even twice, through args parsing), So, be Fast & Clean! This method only reads from the current input stream (Mouth).

pub fn read_x_token(
&mut self,
toplevel_opt: Option<bool>,
commentsok: bool,
state: &mut State
) -> Result<Option<Token>>

pub fn read_raw_line(&mut self, state: &State) -> Option<String>

Read the next raw line (string); primarily to read from the Mouth, but keep any unread input!

pub fn unread(&mut self, tokens: Tokens)

pub fn unread_mut(&mut self, tokens: &mut Tokens)

pub fn unread_one(&mut self, token: Token)

pub fn read_non_space(&mut self, state: &mut State) -> Option<Token>

Read a single non-space token

pub fn read_x_non_space(&mut self, state: &mut State) -> Result<Option<Token>>

Read a single expanded, non-space, token

pub fn read_balanced(
&mut self,
expanded: bool,
state: &mut State
) -> Result<Option<Tokens>>

Read a sequence of tokens balanced in {} assuming the { has already been read. Returns a Tokens list of the balanced sequence, omitting the closing }

pub fn read_keyword(
&mut self,
keywords: &[&str],
state: &mut State
) -> Result<Option<String>>

Match the input against a set of keywords; Similar to readMatch, but the keywords are strings, and Case and catcodes are ignored; additionally, leading spaces are skipped. AND, macros are expanded.

pub fn read_until(&mut self, delim: &Tokens, state: &mut State) -> Result<Tokens>

Return a (balanced) sequence tokens until a match against one of the Tokens in @delims. Note that Braces on input hides the contents from matching, so this assumes there wont be braces in $delim! But, see readUntilBrace for that case.

pub fn read_until_token(&mut self, t: Token, state: &mut State) -> Result<Tokens>

Convenience method wrapping around read_until TODO: This seems to be the wrong Rust type interface, we need to rework…

pub fn read_until_brace(&mut self, state: &mut State) -> Result<Option<Tokens>>

pub fn read_next_conditional(
&mut self,
state: &mut State
) -> Option<(Token, ConditionalType)>

pub fn read_arg(&mut self, state: &mut State) -> Result<Tokens>

Higher-level readers: Read various types of things from the input: tokens, non-expandable tokens, args, Numbers, …

pub fn read_optional(&mut self, state: &mut State) -> Result<Option<Tokens>>

pub fn if_next(&mut self, token: Token, state: &mut State) -> Result<bool>

pub fn read_value(
&mut self,
value_type: RegisterType,
state: &mut State
) -> Result<RegisterValue>

pub fn read_register_value(
&mut self,
value_type: RegisterType,
state: &mut State
) -> Result<Option<RegisterValue>>

pub fn read_match(
&mut self,
choices: &[&Tokens],
state: &mut State
) -> Result<Option<Tokens>>

Match the input against one of the Token or Tokens in @choices; return the matching one or undef.

pub fn read_number(&mut self, state: &mut State) -> Result<Number>

====================================================================== Integer, Number

= = | = |

pub fn read_normal_integer(
&mut self,
state: &mut State
) -> Result<Option<Number>>

= | | ’ | “ | `

pub fn read_float(&mut self, state: &mut State) -> Result<Float>

====================================================================== Float, a floating point number. Similar to factor, but does NOT accept comma! This is NOT part of TeX, but is convenient.

pub fn read_dimension(&mut self, state: &mut State) -> Result<Dimension>

pub fn read_glue(&mut self, state: &mut State) -> Result<Glue>

pub fn read_rubber(
&mut self,
mu: bool,
state: &mut State
) -> Result<(Option<i64>, Option<FillCode>)>

pub fn read_mu_glue(&mut self, state: &mut State) -> Result<MuGlue>

pub fn read_mu_dimension(&mut self, state: &mut State) -> Result<MuDimension>

pub fn read_mu_unit(&mut self, state: &mut State) -> Result<Option<i64>>

pub fn read_tokens_value(&mut self, state: &mut State) -> Result<Tokens>

Apparent behaviour of a token value (ie \toks#=)

pub fn skip_spaces(&mut self, state: &mut State)

pub fn skip_one_space(&mut self, state: &mut State)

pub fn setup_scan(&mut self)

pub fn reading_from_mouth<R, FnR>(
&mut self,
mouth: Mouth,
state: &mut State,
reader: FnR
) -> Rwhere
FnR: FnOnce(&mut Gullet, &mut State) -> R,

pub fn do_expand<T: Into<Tokens>>(
&mut self,
tokens: T,
outer_state: &mut State
) -> Result<Tokens>