Struct rtx_core::keyvals::KeyVals

pub struct KeyVals { /* private fields */ }

Implementations

impl KeyVals

pub fn new(options: KeyvalsConfig) -> Self

====================================================================== The KeyVals constructor

This defines the KeyVals data object that can appear in the datastream along with tokens, boxes, etc. Thus it has to be digestible, however we may not want to digest it more than once.

pub fn get_value(&self, key: &str) -> Option<&ArgWrap>

return the value of a given key. If multiple values are given, return the last one.

pub fn get_value_digested(&self, key: &str) -> Option<&Digested>

return the digested value of a given key. If multiple values are given, return the last one. This call does not digest the value, and will return None if called pre-digestion

pub fn get_values(&self, key: &str) -> Option<&Vec<ArgWrap>>

return a list of values for a given key

pub fn get_pairs(&self) -> Iter<'_, (String, ArgWrap)>

return the set of key-value pairs

pub fn as_flat_hash(self) -> HashMap<String, ArgWrap>

consume KeyVals and return a flat HashMap

pub fn as_hash(self) -> HashMap<String, Vec<ArgWrap>>

consume KeyVals and return the cached HashMap of input values

pub fn as_hash_digested(self) -> HashMap<String, Vec<Digested>>

consume KeyVals and return the cached HashMap of digested values

pub fn get_hash(&self) -> HashMap<String, String>

returns a key => ToString(value)

pub fn get_hash_digested(&self) -> HashMap<String, String>

returns a key => ToString(value)

pub fn get_keyvals(&self) -> &HashMap<String, Vec<ArgWrap>>

pub fn has_key(&self, key: &str) -> bool

pub fn revert(&self) -> Result<Tokens>

pub fn add_value( &mut self, key: &str, value_arg: ArgWrap, use_default: bool, no_rebuild: bool ) -> Result<()>

pub fn set_value( &mut self, key: &str, value: ArgWrap, use_default: bool ) -> Result<()>

pub fn read_from(&mut self, until: Token, silence_missing: bool) -> Result<()>

pub fn into_tokens(self) -> Result<Tokens>

TODO: This is an improvised method for switching KeyVals into Tokens, but losing all collected metadata. The long-term solution ought to be via a type system extension, where the arguments to our before-digest closures are a vector of a new type ReadValue ::= [Token, KeyVals, RegisterValue] potentially? On the other hand, we can also put the extra effort of postponing the build of KV metadata until digestion, this way not losing any time reserializing metadata