Struct rtx_core::state::State

pub struct State {
    pub tag_properties: FxHashMap<SymStr, TagOptions>,
    pub indirect_model: Option<IndirectModel>,
    pub pending_resources: Vec<Resource>,
    pub verbosity: i32,
    pub input_encoding: Option<String>,
    pub search_paths: VecDeque<String>,
    pub graphics_paths: VecDeque<String>,
    pub nomathparse: bool,
    pub bindings_dispatch: Option<BindingDispatcher>,
    pub extra_bindings_dispatch: Option<BindingDispatcher>,
    /* private fields */
}

The state efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

Fields
tag_properties: FxHashMap<SymStr, TagOptions>indirect_model: Option<IndirectModel>

an optional indirect model for long-distance relationships

pending_resources: Vec<Resource>

Document-related resources declared during core conversion, pending until XML is finalized

verbosity: i32input_encoding: Option<String>search_paths: VecDeque<String>

current paths to search for TeX inputs

graphics_paths: VecDeque<String>

current paths to search for graphics

nomathparse: bool

flag to disable math parsing

bindings_dispatch: Option<BindingDispatcher>

A dispatcher routing to the compiled code of the main/official rtx bindings

extra_bindings_dispatch: Option<BindingDispatcher>

Auxiliary convenience – extra dispatch

Implementations

impl State

pub fn new(options: StateOptions) -> Self

pub fn table(&self, name: TableName) -> &Table

borrow/get the named table

pub fn table_mut(&mut self, name: TableName) -> &mut Table

mutably borrow/get the named table

pub fn get_prefix(&self, prefix: &str) -> bool

gets the current value of a named prefix

pub fn assign_value<T: Into<Stored>, S: Into<Option<Scope>>>( &mut self, key: &str, value: T, scope: S, )

assigns a Stored value at the given key and scope

pub fn lookup_value(&self, key: &str) -> Option<&Stored>

fetches a Stored value at the given key, from the Value table

pub fn lookup_value_sym(&self, key: &SymStr) -> Option<&Stored>

pub fn lookup_value_mut<'lv>( &'lv mut self, key: &'lv str, ) -> Option<&mut Stored>

mutably borrows a Stored value at the given key, from the Value table

pub fn lookup_vecdeque<'lvdq>( &'lvdq self, key: &'lvdq str, ) -> Option<&VecDeque<Stored>>

like lookup_value but only recognizes Stored::VecDequeStored

pub fn lookup_font_info(&self, key: &Token) -> Result<Option<&Stored>>

pub fn lookup_mapping(&self, map: &str, key: &str) -> Option<&Stored>

manage a (global) hash of values

pub fn lookup_mapping_keys(&self, map: &str) -> Vec<SymStr>

pub fn lookup_stacked_values(&self, key: &str) -> Vec<&Stored>

pub fn ensure_tag_property(&mut self, tag: SymStr) -> &mut TagOptions