Struct rtx_core::stomach::Stomach

pub struct Stomach {
    pub gullet: Gullet,
    pub token_stack: Vec<Arc<Token>>,
    pub boxing: Vec<Token>,
    pub box_list: Vec<Digested>,
}

Fields
gullet: Gullettoken_stack: Vec<Arc<Token>>boxing: Vec<Token>box_list: Vec<Digested>
Implementations

impl<'t> Stomach

pub fn get_gullet_mut(&mut self) -> &mut Gullet

pub fn get_gullet(&self) -> &Gullet

pub fn get_boxing_level(&self) -> usize

pub fn get_script_level(&self, state: &State) -> usize

ScriptLevel is similar to boxing level, but relative to current Math mode’s level This is used for the scriptpos attribute to recognize overlapping sccripts. Making it relative to the math’s level avoids unnecessary changes

pub fn regurgitate(&mut self) -> Vec<Digested>

pub fn digest_next_body(
&mut self,
terminal_opt: Option<Token>,
state: &mut State
) -> Result<Vec<Digested>>

pub fn digest<T: Into<Tokens>>(
&mut self,
tokens: T,
outer_state: &mut State
) -> Result<Digested>

Digest a list of tokens independent from any current Gullet. Typically used to digest arguments to primitives or constructors. Returns a List containing the digested material.

pub fn raw_tex(&mut self, text: &str, state: &mut State) -> Result<()>

pub fn invoke_token<'a>(
&mut self,
input_token: &'a Token,
state: &mut State
) -> Result<Vec<Digested>>

Invoke a token; If it is a primitive or constructor, the definition will be invoked, possibly arguments will be parsed from the Gullet. Otherwise, the token is simply digested: turned into an appropriate box. Returns a list of boxes/whatsits.

pub fn reading_from_mouth<R, FnR>(
&mut self,
mouth: Mouth,
state: &mut State,
reader: FnR
) -> Rwhere
FnR: FnOnce(&mut Stomach, &mut State) -> R,

pub fn push_stack_frame(&mut self, nobox: bool, state: &mut State)

pub fn pop_stack_frame(&mut self, nobox: bool, state: &mut State) -> Result<()>

pub fn current_frame_message(&self, state: &State) -> String

pub fn bgroup(&mut self, state: &mut State)

if $nobox is true, inhibit incrementing the boxingLevel

pub fn egroup(&mut self, state: &mut State) -> Result<()>

pub fn begingroup(&mut self, state: &mut State)

pub fn endgroup(&mut self, state: &mut State) -> Result<()>

pub fn set_mode(&mut self, mode: &str, state: &mut State) -> Result<()>

This sets the mode without doing any grouping (NOR does it stack the modes!!) Useful for environments, where the group has already been established. (presumably, in the long run, modes & groups should be much less coupled)

pub fn begin_mode(&mut self, mode: &str, state: &mut State) -> Result<()>

pub fn end_mode(&mut self, mode: &str, state: &mut State) -> Result<()>