Module rtx_core::tokens

Grouping Tokens together Token List constructors.

Structs

Tokens

Tokens are a thin wrapper over a vector of Token objects usually read from a Mouth. They are usually treated as an immutable interface, an have to be consumed via .unlist() for access to the underlying data.

Constants

NO_BORROWED_TOKENS

NO_TOKENS

Use this to avoid reallocating a new empty Vec each time you need a placeholder Tokens return value

UNTEX_LINELENGTH

If untex is requested to add line-breaks, this is the line length it will allow