Struct rtx_core::whatsit::Whatsit

pub struct Whatsit {
    pub args: Vec<Option<Digested>>,
    pub properties: HashMap<String, Stored>,
    pub definition: Arc<dyn Definition>,
    pub reversion: Option<Tokens>,
    pub dual_reversion: Option<Tokens>,
    pub locator: Locator,
}

Represents a digested object that can generate arbitrary elements in the XML Document.

Fields
args: Vec<Option<Digested>>

arguments

properties: HashMap<String, Stored>

additional properties, such as font information or sizing

definition: Arc<dyn Definition>

the definition responsible for creating this object

reversion: Option<Tokens>

cached tokens for reverting back (note that the “reversion” property is currently also used)

dual_reversion: Option<Tokens>

special-case reversion tokens for whatsits representing Dual math structures

locator: Locator

point of origin in the source file

Implementations

impl Whatsit

pub fn is_math(&self) -> bool

pub fn is_empty(&self) -> bool

checks if there are any arguments

pub fn set_properties(&mut self, props: HashMap<String, Stored>)

sets a pre-assembled HashMap of properties

pub fn get_definition(&self) -> Arc<dyn Definition>

accessor for the definition which built this Whatsit

pub fn get_arg(&self, n: usize) -> Option<&Digested>

accessor for the argument at index n (starting from 1)

pub fn get_arg_mut(&mut self, n: usize) -> Option<&mut Digested>

mutably borrow the argument at index n (starting from 1)

pub fn get_args(&self) -> &Vec<Option<Digested>>

accessor for the full list of arguments

pub fn set_args(&mut self, args: Vec<Option<Digested>>)

Sets the list of arguments for this whatsit (each arg should be Digested::List).

pub fn get_trailer(&self) -> Option<Digested>

accessor for the trailer property. See whatsit::set_body

pub fn set_body(&mut self, body: Vec<Digested>, state: &mut State)

Sets the body of the whatsit to the boxes in body. The last box in body is assumed to represent the trailer, that is the result of the invocation that closed the environment or math. It is stored separately in the properties under “trailer”.