Module rtx_core::binding::def::dialect

Structs

RegisterOptions

configuration for creating a new Register

Statics

ARG_HOLE

regex for the usual argument placeholders #1-#9

CONDITIONAL_CS_RE

regex for the prefix of a conditional command sequence

DIRTY_ID_IDIOM_RE

regex for ${}^{label}$

HAS_ARG_OR_CS

regex for a TeX argument specifier or any command sequence

LEADING_PROTOCOL_RE

regex for the prefix of a protocol (such as literal:)

NON_ID_CHARSET_RE

regex for characters not expected in a usual id attribute

SPACES_RE

regex for one-or-more spaces

TILDE_NOISE_RE

regex for a strange noisy TeX \\~{}

TRAILING_SLASH_RE

regex for a trailing slash (trivial, but aids replacement of said slash)

Functions

allocate_register

coerce_cs

unconditionally wraps a CS token around a string

def_conditional

Define a conditional control sequence. Its processing takes place in the Gullet. The test is applied to the arguments (if any), which determines which branch is executed. If the test is undefined, the conditional is a “user defined” one; Two additional primitives are defined \footrue and \foofalse; the test is then determined by the most recently called of those.

def_constructor

The Constructor is where LaTeXML really starts getting interesting; invoking the control sequence will generate an arbitrary XML fragment in the document tree. More specifically: during digestion, the arguments will be read and digested, creating a Whatsit to represent the object. During absorption by the Document, the Whatsit will generate the XML fragment according to the compiled_replacement.

def_environment

Defines an Environment that generates a specific XML fragment. compiled_replacement is of the same form as for DefConstructor, but will generally include reference to the #body property. Upon encountering a \begin{env}: the mode is switched, if needed, else a new group is opened; then the environment name is noted; the beforeDigest hook is run. Then the Whatsit representing the begin command (but ultimately the whole environment) is created and the after_digest_begin hook is run. Next, the body will be digested and collected until the balancing \end{env}. Then, any after_digest hook is run, the environment is ended, finally the mode is ended or the group is closed. The body and \end{env} whatsit are added to the \begin{env}’s whatsit as body and trailer, respectively.

def_macro

Defines the macro expansion for a cs: a macro control sequence that reads parameters as specified by paramlist and is expanded during macro expansion time in the Gullet. See ExpansionBody for the possible kinds of expansion material.

def_math

Define a Mathematical symbol or function. There are two sets of cases: (1) If the presentation appears to be TeX code, we create an XMDual, since the presentation may end up with structure, etc. (2) But if the presentation is a simple string, or unicode, it is just the content of the symbol; even if the function takes arguments.

def_math_constructor

Uses of DefMath without arguments, but with constructor-like options, are realized via a Constructor definition

def_math_dual

Advanced math replacements require a XMDual representation

def_math_primitive

EXPERIMENT: Introduce an intermediate case for simple symbols Define a primitive that will create a Tbox with the appropriate set of XMTok attributes.

def_primitive

Defines a primitive control sequence; a primitive is processed during digestion (in the Stomach), after macro expansion but before Construction time. Primitive control sequences generate Boxes or Lists, generally containing basic Unicode content, rather than structured XML. Primitive control sequences are also executed for side effect during digestion, effecting changes to the State.

def_register

Defines a register with value as the initial value (a Number, Dimension, Glue, MuGlue or Tokens — I haven’t handled Box’s yet). Usually, the prototype is just the control sequence, but registers are also handled by prototypes like \count{Number}. DefRegister arranges that the register value can be accessed when a numeric, dimension, … value is being read, and also defines the control sequence for assignment.

dualize_arglist

Given a list of Tokens (to be expanded into mathematical objects) return two lists: (1) The Tokens’ wrapped in an XMAarg, with an ID added (2) a corresponding list of Tokens creating XMRef’s to those IDs

get_xmarg_id

is_definable

Check if the token is not yet defined, or let to \relax

is_defined

Is defined in the LaTeX-y sense of also not being let to \relax.

is_defined_token

Token variant of is_defined. Defined in the LaTeX-y sense of also not being let to \relax.