Module rtx_core::state

A global, singleton, mutable State - hosts almost all TeX-facing runtime information for the conversion

Re-exports

pub use crate::common::store::Stored;

Structs

State

The State efficiently maintain the bindings in a TeX-like fashion. bindings associate data with keys (eg definitions with macro names) and respect TeX grouping; that is, an assignment is only in effect until the current group (opened by \bgroup) is closed (by \egroup).

StateOptions

State fields allowed for customization during construction

UndoFrame

For each of several tables (being “value”, “meaning”, “catcode” or other space of names), each table maintains the bound values, and “undo” defines the stack frames

Enums

Catcodes

High-level catcode profiles

Scope

installation scope in the State tables

TableName

the kinds of tables bookkept in the State

Statics

STD_STATE

STY_STATE

UNITS

Used in conversion to scaled points

Type Definitions

AssignmentCount

Ledger for stacked assignments

StashTable

The (table_name, key, value) contents of a stored table of assignments

Table

There are tables for catcode: keys are char; Also, math:char =1 when char is active in math. mathcode, sfcode, lccode, uccode, delcode : are similar to catcode but store additional kinds codes per char (see TeX) value: keys are anything (typically a string, though) and value is the value associated with it meaning: The definition assocated with key, usually a control-sequence. stash & stash_active: support named scopes (see also activateScope & deactivateScope)