rtx_core
Module stomach

The stomach is an abstraction responsible for digesting Tokens and Registers prepared by the Gullet into Boxes

Structs

Stomach

The Stomach is responsible for digesting tokens into boxes, lists, etc.

Statics

STOMACH

Functions

begin_mode

Begin processing in mode; one of “text”, “display-math” or “inline-math”. This also begins a new level of grouping and switches to a font appropriate for the mode.

begingroup

Begin a new level of binding by pushing a new stack frame.

bgroup

Begin a new level of binding by pushing a new stack frame, and a new level of boxing the digested output.

clone_box_list

current_frame_message

explain the current frame

digest

Digest a list of tokens independent from any current Gullet. Typically used to digest arguments to primitives or constructors. Returns a List containing the digested material.

digest_next_body

Return the digested List after reading and digesting a body from the its Gullet. The body extends until the current level of boxing or environment is closed.

egroup

End a level of binding by popping the last stack frame, undoing whatever bindings appeared there, and also decrementing the level of boxing.

end_mode

End processing in mode; an error is signalled if stomach is not currently in mode. This also ends a level of grouping.

endgroup

End a level of binding by popping the last stack frame, undoing whatever bindings appeared there.

expire_local_box_list

extend_box_list

get_boxing_level

get the current boxing level

get_script_level

ScriptLevel is similar to boxing level, but relative to current Math mode’s level

initialize_stomach

Initialize various stomach parameters, preload, etc.

invoke_token

Invoke a token

new_local_box_list

pop_box_list

pop_stack_frame

Removes the last/current stack frame, ending a TeX group

push_box_list

push_stack_frame

Adds a new stack frame for a TeX group.

raw_tex

a convenience function for including chunks of raw TeX (or LaTeX) code It is useful for copying portions of the normal implementation that can be handled simply using macros and primitives.

regurgitate

steal the previously digested boxes from the current level.

set_mode

Sets the mode without doing any grouping (NOR does it stack the modes!!)

set_stomach