Trait rtx_core::BoxOps

pub trait BoxOps: Object {

    // Required methods
    fn be_absorbed(
        &self,
        document: &mut Document,
        state: &mut State
    ) -> Result<Vec<Node>>;
    fn get_string(&self, state: &State) -> Result<Cow<'_, str>>;
    fn with_properties<R, FnR>(&self, caller: FnR) -> R
       where FnR: FnOnce(&HashMap<String, Stored>) -> R;
    fn get_font(&self, state: &mut State) -> Result<Option<Cow<'_, Font>>>;
    fn compute_size(
        &self,
        options: HashMap<String, Stored>,
        state: &mut State
    ) -> Result<(Dimension, Dimension, Dimension)>;

    // Provided methods
    fn unlist(&self) -> Vec<Digested> { ... }
    fn unlist_ref(&self) -> Vec<Cow<'_, Digested>> { ... }
    fn be_absorbed_mut(
        &mut self,
        _document: &mut Document,
        _state: &mut State
    ) -> Result<Vec<Node>> { ... }
    fn get_tokens(&self) -> Option<&Tokens> { ... }
    fn get_properties(&self) -> &HashMap<String, Stored> { ... }
    fn get_properties_mut(&mut self) -> &mut HashMap<String, Stored> { ... }
    fn set_property<T: Into<Stored>>(&mut self, key: &str, value: T) { ... }
    fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>> { ... }
    fn get_property_string(&self, key: &str) -> String { ... }
    fn get_property_mut(&mut self, key: &str) -> Option<&mut Stored> { ... }
    fn has_property(&self, key: &str) -> bool { ... }
    fn get_property_bool(&self, key: &str) -> bool { ... }
    fn get_body(&self) -> Result<Option<Digested>> { ... }
    fn set_font(&mut self, _font: Rc<Font>) { ... }
    fn set_width<T: Into<Stored>>(&mut self, width: T) { ... }
    fn get_width(
        &self,
        options: Option<HashMap<String, Stored>>,
        state: &mut State
    ) -> Result<Option<RegisterValue>> { ... }
    fn set_height<T: Into<Stored>>(&mut self, width: T) { ... }
    fn get_height(&self, _state: &State) -> Option<RegisterValue> { ... }
    fn set_depth<T: Into<Stored>>(&mut self, width: T) { ... }
    fn get_depth(&self, _state: &State) -> Option<RegisterValue> { ... }
    fn get_size(
        &mut self,
        options: Option<HashMap<String, Stored>>,
        state: &mut State
    ) -> Result<(Dimension, Dimension, Dimension, Dimension, Dimension, Dimension)> { ... }
    fn compute_size_and_cache(
        &mut self,
        options: HashMap<String, Stored>,
        state: &mut State
    ) -> Result<(Dimension, Dimension, Dimension)> { ... }

}

Common operations for Box-like (digested) data

Required Methods

fn be_absorbed( &self, document: &mut Document, state: &mut State ) -> Result<Vec<Node>>

absorb the current object into the Document XML - returning the corresponding nodes

fn get_string(&self, state: &State) -> Result<Cow<'_, str>>

build a string representation of the underlying digested data

fn with_properties<R, FnR>(&self, caller: FnR) -> Rwhere FnR: FnOnce(&HashMap<String, Stored>) -> R,

execute a function using this object’s named properties

fn get_font(&self, state: &mut State) -> Result<Option<Cow<'_, Font>>>

gets the associated font, if any

fn compute_size( &self, options: HashMap<String, Stored>, state: &mut State ) -> Result<(Dimension, Dimension, Dimension)>

computes and returns the size of a box-like object

Provided Methods

fn unlist(&self) -> Vec<Digested>

If composite, unwrap into the contained digested objects (or return self)

fn unlist_ref(&self) -> Vec<Cow<'_, Digested>>

fn be_absorbed_mut( &mut self, _document: &mut Document, _state: &mut State ) -> Result<Vec<Node>>

be_absorbed but with allowed side-effects on the carrier (for Alignment only)

fn get_tokens(&self) -> Option<&Tokens>

get the underlying tokens (preceding digestion)

fn get_properties(&self) -> &HashMap<String, Stored>

deprecated: get the map of named properties. This can not be usable as long as we have any data behind a RefCell wrapper. Use with_properties instead.

fn get_properties_mut(&mut self) -> &mut HashMap<String, Stored>

get a mutable reference to the map of named properties

fn set_property<T: Into<Stored>>(&mut self, key: &str, value: T)

set a named property (allows all Stored types for values)

fn get_property(&self, key: &str) -> Option<Cow<'_, Stored>>

get a single named property (with special “isSpace” check)

fn get_property_string(&self, key: &str) -> String

fn get_property_mut(&mut self, key: &str) -> Option<&mut Stored>

get a mutable reference to a single named property (does NOT have the “isSpace” check)

fn has_property(&self, key: &str) -> bool

checks if a property key has been set

fn get_property_bool(&self, key: &str) -> bool

obtains a boolean property value (false unless Stored::Bool)

fn get_body(&self) -> Result<Option<Digested>>

obtains the “body” of a digested object which captured it

fn set_font(&mut self, _font: Rc<Font>)

sets an associated font

fn set_width<T: Into<Stored>>(&mut self, width: T)

sets a “width” property, for sizing

fn get_width( &self, options: Option<HashMap<String, Stored>>, state: &mut State ) -> Result<Option<RegisterValue>>

gets the “width” property value, if any

fn set_height<T: Into<Stored>>(&mut self, width: T)

sets a “height” property value, for sizing

fn get_height(&self, _state: &State) -> Option<RegisterValue>

gets the “height” property value, if any

fn set_depth<T: Into<Stored>>(&mut self, width: T)

sets a “depth” property value, for sizing

fn get_depth(&self, _state: &State) -> Option<RegisterValue>

gets the “depth” property value, if any

fn get_size( &mut self, options: Option<HashMap<String, Stored>>, state: &mut State ) -> Result<(Dimension, Dimension, Dimension, Dimension, Dimension, Dimension)>

gets the box size as a triple of (width, height, depth) the generic implementation is immutable and will recompute the size on each call see Digested::get_size for a variant with interior mutability which caches the box size

fn compute_size_and_cache( &mut self, options: HashMap<String, Stored>, state: &mut State ) -> Result<(Dimension, Dimension, Dimension)>

computes and caches (via named properties) the size of a box-like object